requirements.txt
opencv-contrib-python
numpy
torch
ultralytics
mediapipe
scikit-learn
matplotlib
PyYAML
streamlit
youtube-dl
google-generativeai
python-dotenv
deepface